fix(quiz): debugging and fixBug

+++ src/controller/quiz-controller.ts
@@ -17,28 +17,53 @@ export const getAllQuiz = async (request: Request, response: Response) => {
             }  
         })
 
-        return response.status(200).json({
+            response.status(200).json({
             success: true,
             data : AllQuiz,
             message : "All quiz found successfully"
         })
+        return
     } catch (error) {
-        return response.status(400).json({
-            success: false,
-            message: `failed to fetch all quiz. ${error}`
+            console.error(error)
+            response.status(500).json({
+              success: false,
+              message: `failed to fetch all quiz.`
         })
+        return
     }
 };
 
 export const getQuizById = async (request: Request, response: Response) => {
     try {
         const idQuiz = request.params.idQuiz;
+        const id = Number(idQuiz)
 
         if (!idQuiz) {
-            return response.status(400).json({
+                response.status(400).json({
                 success: false,
                 message: "id Quiz is required"
             })
+            return
+        }
+
+        if (Number.isNaN(id)) {{
+            response.status(400).json({
+                success: false,
+                message: "id Quiz must be a number"
+            })
+            return
+        }}
+
+        const findQuiz = await prisma.quiz.findFirst({
+          where: { idQuiz: id }
+        })
+
+        if (!findQuiz) {
+            response.status(404).json({
+                success: false,
+                message: "quiz not found"
+            })
+            return
         }
 
         const AllQuiz = await prisma.quiz.findUnique({
@@ -51,24 +76,20 @@ export const getQuizById = async (request: Request, response: Response) => {
 
         })
 
-        if (!AllQuiz) {
-            return response.status(404).json({
-                success: false,
-                message: "quiz not found"
-            })
-        }
-
-        return response.status(200).json({
+          response.status(200).json({
             success: true,
             data : AllQuiz,
             message : "quiz found successfully"
         })
+        return
 
     } catch (error) {
-        return response.status(400).json({
-            success: false,
-            message: `failed to fetch quiz. ${error}`
+            console.error(error)
+            response.status(500).json({
+              success: false,
+              message: `failed to fetch quiz.`
         })
+        return
     }
 }
 
@@ -79,10 +100,11 @@ export const createQuiz = async (request: Request, response: Response) => {
       const uuid = uuidv4();
   
       if (!duration || duration <= 0) {
-        return response.status(400).json({
+          response.status(400).json({
           success: false,
           message: "duration must be greater than 0"
         })
+        return
       }
     
         const newQuiz = await prisma.quiz.create({
@@ -98,25 +120,48 @@ export const createQuiz = async (request: Request, response: Response) => {
           }
         });
     
-        return response.status(200).json({
-          status: true,
+          response.status(201).json({
+          success: true,
           data: newQuiz
         })
+        return
     
       } catch (error) {
-        console.error(error);
-  
-        return response.status(500).json({
-          status: false,
-          message: `Failed to create quiz${error}`,
+          console.error(error);
+          response.status(500).json({
+            success: false,
+            message: `Failed to create quiz.`,
         })
+        return
       }
 }
+
 
 export const updateQuiz = async (request: Request, response: Response) => {
     try {
         const { idQuiz } = request.params;
         const { quiz_title, quiz_date, duration, status, difficulty } = request.body;
+        const id = Number(idQuiz)
+
+        if (Number.isNaN(id)) {{
+          response.status(400).json({
+              success: false,
+              message: "id Quiz must be a number"
+          })
+          return
+      }}
+
+      const findQuiz = await prisma.quiz.findFirst({
+        where: { idQuiz: id }
+      })
+
+      if (!findQuiz) {
+          response.status(404).json({
+              success: false,
+              message: "quiz not found"
+          })
+          return
+      }
     
         const updatedQuiz = await prisma.quiz.update({
           where: { idQuiz: Number(idQuiz) },
@@ -129,17 +174,20 @@ export const updateQuiz = async (request: Request, response: Response) => {
           }
         });
     
-        return response.status(200).json({
+          response.status(200).json({
           status: true,
           data: updatedQuiz,
           message: "Quiz updated successfully"
         })
+        return
     
       } catch (error) {
-        return response.status(500).json({
-          status: false,
-          message: `Failed to update quiz. ${error}`,
+          console.error(error)
+          response.status(500).json({
+            status: false,
+            message: `Failed to update quiz.`,
         })
+        return
       }
 }
 
@@ -147,21 +195,45 @@ export const updateQuiz = async (request: Request, response: Response) => {
 export const deleteQuiz = async (request: Request, response: Response) => {
     try {
       const { idQuiz } = request.params;
+      const id = Number(idQuiz)
+
+      if (Number.isNaN(id)) {
+        response.status(400).json({
+            success: false,
+            message: "id Quiz must be a number"
+        })
+        return
+    }
+
+    const findQuiz = await prisma.quiz.findFirst({
+      where: { idQuiz: id }
+    })
+
+    if (!findQuiz) {
+        response.status(404).json({
+            success: false,
+            message: "quiz not found"
+        })
+        return
+    }
   
       const deletedQuiz = await prisma.quiz.delete({
         where: { idQuiz: Number(idQuiz) }
       });
   
-      return response.status(200).json({
+        response.status(200).json({
         status: true,
         data: deletedQuiz,
         message: "Quiz deleted successfully"
       })
+      return
   
     } catch (error) {
-      return response.status(500).json({
-        status: false,
-        message: `Failed to delete quiz. ${error}`,
+      console.error(error)
+        response.status(500).json({
+          status: false,
+          message: `Failed to delete quiz.`,
       })
+      return
     }
 }
